extract test data factory to separate file
