+++ src/NPetrovich.Tests/NPetrovich.Tests.csproj
@@ -53,6 +53,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Fixtures\GenderToolsFixture.cs" />
     <Compile Include="Fixtures\CaseInflectionFixture.cs" />
+    <Compile Include="TestDataProviders\InflectionTestCaseDataFactory.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Data\FirstNames.csv">
